@@ -13,6 +13,8 @@ use Grav\Common\Grav;
 use Grav\Common\Session;
 use Grav\Common\User\Interfaces\UserCollectionInterface;
 use Grav\Common\User\Interfaces\UserInterface;
+use Grav\Framework\Session\SessionInterface;
+use Grav\Plugin\Login\Login;
 use RocketTheme\Toolbox\Event\Event;
 
 /**
@@ -33,37 +35,42 @@ class UserLoginEvent extends Event
     /**
      * Undefined event state.
      */
-    const AUTHENTICATION_UNDEFINED = 0;
+    public const AUTHENTICATION_UNDEFINED = 0;
 
     /**
      * onUserAuthenticate success.
      */
-    const AUTHENTICATION_SUCCESS = 1;
+    public const AUTHENTICATION_SUCCESS = 1;
 
     /**
      * onUserAuthenticate fails on bad username/password.
      */
-    const AUTHENTICATION_FAILURE = 2;
+    public const AUTHENTICATION_FAILURE = 2;
 
     /**
      * onUserAuthenticate fails on auth cancellation.
      */
-    const AUTHENTICATION_CANCELLED = 4;
+    public const AUTHENTICATION_CANCELLED = 4;
 
     /**
      * onUserAuthorizeLogin fails on expired account.
      */
-    const AUTHORIZATION_EXPIRED = 8;
+    public const AUTHORIZATION_EXPIRED = 8;
 
     /**
-     * onUserAuthorizeLogin is delayed until user has performed extra action(s).
+     * onUserAuthorizeLogin is delayed until user has performed AUTHORIZATION_CHALLENGE.
      */
-    const AUTHORIZATION_DELAYED = 16;
+    public const AUTHORIZATION_DELAYED = 16;
 
     /**
      * onUserAuthorizeLogin fails for other reasons.
      */
-    const AUTHORIZATION_DENIED = 32;
+    public const AUTHORIZATION_DENIED = 32;
+
+    /**
+     * onUserAuthorizeLogin was challenged, combine with AUTHENTICATION_SUCCESS, AUTHENTICATION_FAILURE or AUTHENTICATION_CANCELLED.
+     */
+    public const AUTHORIZATION_CHALLENGE = 64;
 
     /**
      * UserLoginEvent constructor.
@@ -92,11 +99,24 @@ class UserLoginEvent extends Event
         if (!$this->offsetExists('user')) {
             /** @var UserCollectionInterface $users */
             $users = Grav::instance()['accounts'];
-            $this->offsetSet('user', $users->load($this['credentials']['username']));
+            $user = $users->load($this['credentials']['username']);
+            $this->offsetSet('user', $user);
+
+            if (Login::DEBUG) {
+                if ($user->exists()) {
+                    Login::addDebugMessage('Login user:', $user);
+                } else {
+                    Login::addDebugMessage("Login: user '{$this['credentials']['username']}' not found");
+                }
+            }
+
         }
     }
 
-    public function isSuccess()
+    /**
+     * @return bool
+     */
+    public function isSuccess(): bool
     {
         $status = $this->offsetGet('status');
         $failure = static::AUTHENTICATION_FAILURE | static::AUTHENTICATION_CANCELLED | static::AUTHORIZATION_EXPIRED
@@ -105,15 +125,28 @@ class UserLoginEvent extends Event
         return ($status & static::AUTHENTICATION_SUCCESS) && !($status & $failure);
     }
 
-    public function isDelayed()
+    /**
+     * @return bool
+     */
+    public function isDelayed(): bool
     {
         return $this->isSuccess() && ($this->offsetGet('status') & static::AUTHORIZATION_DELAYED);
     }
 
     /**
+     * @return bool
+     */
+    public function isChallenged(): bool
+    {
+        $status = $this->offsetGet('status');
+
+        return (bool)($status & static::AUTHORIZATION_CHALLENGE);
+    }
+
+    /**
      * @return int
      */
-    public function getStatus()
+    public function getStatus(): int
     {
         return (int)$this->offsetGet('status');
     }
@@ -122,7 +155,7 @@ class UserLoginEvent extends Event
      * @param int $status
      * @return $this
      */
-    public function setStatus($status)
+    public function setStatus($status): self
     {
         $this->offsetSet('status', $this->offsetGet('status') | (int)$status);
 
@@ -132,7 +165,7 @@ class UserLoginEvent extends Event
     /**
      * @return array
      */
-    public function getCredentials()
+    public function getCredentials(): array
     {
         return $this->offsetGet('credentials') + ['username' => '', 'password' => ''];
     }
@@ -151,7 +184,7 @@ class UserLoginEvent extends Event
      * @param mixed $value
      * @return $this
      */
-    public function setCredential($name, $value)
+    public function setCredential($name, $value): self
     {
         $this->items['credentials'][$name] = $value;
 
@@ -161,7 +194,7 @@ class UserLoginEvent extends Event
     /**
      * @return array
      */
-    public function getOptions()
+    public function getOptions(): array
     {
         return $this->offsetGet('options');
     }
@@ -180,7 +213,7 @@ class UserLoginEvent extends Event
      * @param mixed $value
      * @return $this
      */
-    public function setOption($name, $value)
+    public function setOption($name, $value): self
     {
         $this->items['options'][$name] = $value;
 
@@ -188,9 +221,9 @@ class UserLoginEvent extends Event
     }
 
     /**
-     * @return Session|null
+     * @return SessionInterface|Session|null
      */
-    public function getSession()
+    public function getSession(): ?SessionInterface
     {
         return $this->offsetGet('session');
     }
@@ -198,7 +231,7 @@ class UserLoginEvent extends Event
     /**
      * @return UserInterface
      */
-    public function getUser()
+    public function getUser(): UserInterface
     {
         return $this->offsetGet('user');
     }
@@ -207,7 +240,7 @@ class UserLoginEvent extends Event
      * @param UserInterface $user
      * @return $this
      */
-    public function setUser(UserInterface $user)
+    public function setUser(UserInterface $user): self
     {
         $this->offsetSet('user', $user);
 
@@ -217,7 +250,7 @@ class UserLoginEvent extends Event
     /**
      * @return array
      */
-    public function getAuthorize()
+    public function getAuthorize(): array
     {
         return (array)$this->offsetGet('authorize');
     }
@@ -225,15 +258,15 @@ class UserLoginEvent extends Event
     /**
      * @return string|null
      */
-    public function getMessage()
+    public function getMessage(): ?string
     {
         return !empty($this->items['message'][0]) ? (string)$this->items['message'][0] : null;
     }
 
     /**
-     * @return string|null
+     * @return string
      */
-    public function getMessageType()
+    public function getMessageType(): string
     {
         return !empty($this->items['message'][1]) ? (string)$this->items['message'][1] : 'info';
     }
@@ -243,7 +276,7 @@ class UserLoginEvent extends Event
      * @param string|null $type
      * @return $this
      */
-    public function setMessage($message, $type = null)
+    public function setMessage($message, $type = null): self
     {
         $this->items['message'] = $message ? [$message, $type] : null;
 
@@ -255,7 +288,7 @@ class UserLoginEvent extends Event
      * @param string|null $type
      * @return $this
      */
-    public function defMessage($message, $type = null)
+    public function defMessage($message, $type = null): self
     {
         if ($message && !isset($this->items['message'])) {
             $this->setMessage($message, $type);
@@ -267,7 +300,7 @@ class UserLoginEvent extends Event
     /**
      * @return string|null
      */
-    public function getRedirect()
+    public function getRedirect(): ?string
     {
         return $this->items['redirect'] ?? null;
     }
@@ -275,9 +308,9 @@ class UserLoginEvent extends Event
     /**
      * @return int
      */
-    public function getRedirectCode()
+    public function getRedirectCode(): int
     {
-        return $this->items['redirect_code'] ?? 303;
+        return (int)($this->items['redirect_code'] ?? 303);
     }
 
     /**
@@ -285,7 +318,7 @@ class UserLoginEvent extends Event
      * @param int $code
      * @return $this
      */
-    public function setRedirect($path, $code = 303)
+    public function setRedirect($path, $code = 303): self
     {
         $this->items['redirect'] = $path ?: null;
         $this->items['redirect_code'] = (int)$code;
@@ -298,7 +331,7 @@ class UserLoginEvent extends Event
      * @param int $code
      * @return $this
      */
-    public function defRedirect($path, $code = 303)
+    public function defRedirect($path, $code = 303): self
     {
         if ($path && !isset($this->items['redirect'])) {
             $this->setRedirect($path, $code);
@@ -313,7 +346,7 @@ class UserLoginEvent extends Event
      * @param mixed $offset Asset name value
      * @param mixed $value  Asset value
      */
-    public function __set($offset, $value)
+    public function __set($offset, $value): void
     {
         $this->offsetSet($offset, $value);
     }
@@ -335,7 +368,7 @@ class UserLoginEvent extends Event
      * @param  mixed   $offset Asset name value
      * @return boolean         True if the value is set
      */
-    public function __isset($offset)
+    public function __isset($offset): bool
     {
         return $this->offsetExists($offset);
     }
@@ -345,8 +378,16 @@ class UserLoginEvent extends Event
      *
      * @param mixed $offset The name value to unset
      */
-    public function __unset($offset)
+    public function __unset($offset): void
     {
         $this->offsetUnset($offset);
+    }
+
+    /**
+     * @return array
+     */
+    public function __debugInfo(): array
+    {
+        return get_object_vars($this);
     }
 }
