@@ -26,11 +26,13 @@ class MapMarkerLeafletPlugin extends Plugin
         ]);
         //add assets
         $assets = $this->grav['assets'];
-        $assets->addJs("https://unpkg.com/leaflet@1.3.4/dist/leaflet.js");
-        $assets->addCss("https://unpkg.com/leaflet@1.3.4/dist/leaflet.css");
-        //add leaflet awesome assets
-        $assets->addJs('plugin://map-marker-leaflet/assets/leaflet.awesome-markers.js');
-        $assets->addCss('plugin://map-marker-leaflet/assets/leaflet.awesome-markers.css');
+
+        $assets->registerCollection('leaflet', [
+            'plugin://map-marker-leaflet/assets/leaflet.js',
+            'plugin://map-marker-leaflet/assets/leaflet.css'
+        ]);
+
+        $assets->add('leaflet', 90);
     }
 
     public function onTwigTemplatePaths()
